extracting landing screen widgets and setup in common widgets
